pyproject.toml
[tool.poetry]
name = "g-kingdom"
version = "0.1.0"
description = ""
authors = ["Alexander Kupreyeu <agkupreev@gmail.com>"]
maintainers = ["Alexander Kupreyeu <agkupreev@gmail.com>"]
license = "MIT"
readme = "README.rst"
homepage = ""
repository = ""
documentation = ""
keywords = ["Flask",
            "Poetry",
            "API",
            "Game"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
flask-restful = "^0.3.8"
flask-migrate = "^2.5.3"
flask-sqlalchemy = "^2.4.1"
marshmallow-sqlalchemy = "^0.22.3"
marshmallow = "^3.5.1"
flask-marshmallow = "^0.11.0"
flask-jwt-extended = "^3.24.1"
passlib = "^1.7.2"
python-dotenv = "^0.13.0"
apispec = {extras = ["yaml"], version = "^3.3.0"}
simplejson = "^3.17.0"
apispec-webframeworks = "^0.5.2"
celery = {extras = ["redis"], version = "^4.4.2"}
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
tox = "^3.14.6"
factory_boy = "^2.12.0"
pytest = "^5.4"
pytest-factoryboy = "^2.0.3"
flake8 = "^3.7.9"
black = "^19.10b0"

[tool.poetry.scripts]
manage = "app.manage:cli"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

